requirements.txt
anyascii==0.3.0
asgiref==3.5.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.1
Django==3.2.12
django-filter==21.1
django-modelcluster==5.2
django-otp==1.1.3
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
openpyxl==3.0.9
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
telepath==0.2
typing_extensions==4.1.1
urllib3==1.26.8
wagtail==2.16.1
webencodings==0.5.1
Willow==1.4.1
wincertstore==0.2
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
